screenのパッケージ構成変更
