+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/favorite/FavoriteViewModel.kt
@@ -1,7 +1,7 @@
 package com.sw.sw_api_kotlin_project.screen.favorite
 
 import androidx.lifecycle.liveData
-import com.sw.sw_api_kotlin_project.base.BaseViewModel
+import com.sw.sw_api_kotlin_project.screen.base.BaseViewModel
 import com.sw.sw_api_kotlin_project.repository.FavoriteRepository
 import com.sw.sw_api_kotlin_project.utils.Resource
 import dagger.hilt.android.lifecycle.HiltViewModel
