+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/people/detail/PeopleDetailsViewModel.kt
@@ -1,11 +1,11 @@
-package com.sw.sw_api_kotlin_project.screen.film_details
+package com.sw.sw_api_kotlin_project.screen.people.detail
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import com.sw.sw_api_kotlin_project.base.BaseViewModel
+import com.sw.sw_api_kotlin_project.screen.base.BaseViewModel
 import com.sw.sw_api_kotlin_project.data.database.Favorite
-import com.sw.sw_api_kotlin_project.data.model.Film
+import com.sw.sw_api_kotlin_project.data.model.People
 import com.sw.sw_api_kotlin_project.repository.FavoriteRepository
 import com.sw.sw_api_kotlin_project.utils.DateFormatter
 import com.sw.sw_api_kotlin_project.utils.ListType
@@ -14,7 +14,7 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class FilmDetailsViewModel @Inject constructor(
+class PeopleDetailsViewModel @Inject constructor(
     private val favoriteRepository: FavoriteRepository
 ) : BaseViewModel() {
     private val _favoriteStatus = MutableLiveData<Boolean>()
@@ -26,17 +26,17 @@ class FilmDetailsViewModel @Inject constructor(
         }
     }
 
-    fun addOrDeleteFavorite(film: Film) {
+    fun addOrDeleteFavorite(people: People) {
         viewModelScope.launch {
-            val favorite: Favorite? = favoriteRepository.getFavoriteState(film.title)
+            val favorite: Favorite? = favoriteRepository.getFavoriteState(people.name)
             if (favorite == null) {
                 favoriteRepository.insert(
                     Favorite(
                         id = 0,
-                        name = film.title,
-                        listType = ListType.FILM,
-                        people = null,
-                        film = film,
+                        name = people.name,
+                        listType = ListType.PEOPLE,
+                        people = people,
+                        film = null,
                         planet = null,
                         registrationDate = DateFormatter.getTodayDateStringYYYYMMDDHHMMSS()
                     )
@@ -44,7 +44,7 @@ class FilmDetailsViewModel @Inject constructor(
             } else {
                 favoriteRepository.delete(favorite)
             }
-            getFavoriteState(film.title)
+            getFavoriteState(people.name)
         }
     }
