+++ SW_API_Kotlin_Project/app/src/main/res/layout/fragment_film_details.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.sw.sw_api_kotlin_project.screen.film_details.FilmDetailsViewModel" />
+            type="com.sw.sw_api_kotlin_project.screen.film.detail.FilmDetailsViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
