+++ SW_API_Kotlin_Project/app/src/main/res/layout/fragment_people_details.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.sw.sw_api_kotlin_project.screen.people_details.PeopleDetailsViewModel" />
+            type="com.sw.sw_api_kotlin_project.screen.people.detail.PeopleDetailsViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
