+++ SW_API_Kotlin_Project/app/src/main/res/navigation/nav_film_graph.xml
@@ -17,7 +17,7 @@
 
     <fragment
         android:id="@+id/nav_films_detail"
-        android:name="com.sw.sw_api_kotlin_project.screen.film_details.FilmsDetailsFragment"
+        android:name="com.sw.sw_api_kotlin_project.screen.film.detail.FilmDetailsFragment"
         android:label="FilmsDetailsFragment"
         tools:layout="@layout/fragment_film_details">
         <argument
