+++ SW_API_Kotlin_Project/app/src/main/res/navigation/nav_planet_graph.xml
@@ -17,7 +17,7 @@
 
     <fragment
         android:id="@+id/nav_planet_detail"
-        android:name="com.sw.sw_api_kotlin_project.screen.planet_details.PlanetDetailsFragment"
+        android:name="com.sw.sw_api_kotlin_project.screen.planet.details.PlanetDetailsFragment"
         android:label="PlanetDetailsFragment"
         tools:layout="@layout/fragment_planet_details">
         <argument
